requirements-dev.txt
-r requirements.txt
black>=23.3.0
flake8>=6.0.0
pytest>=7.3.1
pytest-django>=4.5.2
django-debug-toolbar>=4.1.0 